pyproject.toml
[tool.poetry]
name = "fetap"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"  # To satisfy statemachine's requirement
click = "^8.1.7"
python-statemachine = "^2.1.2"
rpi-gpio = {version="==0.7.1", markers="platform_machine == 'aarch64'"}
requests = "^2.31.0"
flask = "^3.0.3"
filelock = "^3.13.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
python-statemachine = {extras = ["diagrams"], version = "^2.1.2"}
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
